# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Piano Shop Api",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/register_user": {
            "post": {
                "tags": [
                    "Auth Api"
                ],
                "description": "use for register user to database",
                "operationId": "4b0da3d32d94b5863496e2f1d0b49060",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Enter your name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Enter your email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Enter your password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/login_user": {
            "post": {
                "tags": [
                    "Auth Api"
                ],
                "description": "use for login user ",
                "operationId": "8d024b1de171512d2fba029968905c03",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Enter your email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Enter your password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/user_token_logout/{token}": {
            "post": {
                "tags": [
                    "Auth Api"
                ],
                "description": "use for logout user by token",
                "operationId": "bd10e2acea589108affe46072d7ce6ed",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Enter your token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/user_token/{token}": {
            "get": {
                "tags": [
                    "Auth Api"
                ],
                "description": "use for get user by token",
                "operationId": "489470ed044137cf55a7ef2d1feed823",
                "parameters": [
                    {
                        "name": "token",
                        "in": "path",
                        "description": "Enter your token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/upload_file": {
            "post": {
                "tags": [
                    "Upload File Api"
                ],
                "description": "use for upload a file to public folder",
                "operationId": "a9ad685a886801e3146628e24beea75f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "add your file",
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Its Ok",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/delete_file/{fileName}": {
            "delete": {
                "tags": [
                    "Upload File Api"
                ],
                "description": "use for delete a file from public folder",
                "operationId": "b27c1ab4bf84efc20437682cf5ec5beb",
                "parameters": [
                    {
                        "name": "fileName",
                        "in": "path",
                        "description": "The name of the file to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File deleted successfully",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    }
}